requirements.txt
Django==1.6.5
argparse==1.2.1
django-bootstrap3==4.6.0
django-haystack==2.1.0
django-registration==1.0
django-widget-tweaks==1.3
psycopg2==2.5.3
pyelasticsearch==0.6.1
requests==2.3.0
simplejson==3.5.2
six==1.6.1
wsgiref==0.1.2
